+++ tests/serialization/test_parameter_serialization.py
@@ -9,25 +9,25 @@ from baybe.parameters.base import Parameter
 
 from ..conftest import _CHEM_INSTALLED
 from ..hypothesis_strategies.parameters import (
-    categorical_parameter,
-    custom_parameter,
-    numerical_continuous_parameter,
-    numerical_discrete_parameter,
-    substance_parameter,
-    task_parameter,
+    categorical_parameters,
+    custom_parameters,
+    numerical_continuous_parameters,
+    numerical_discrete_parameters,
+    substance_parameters,
+    task_parameters,
 )
 
 
 @pytest.mark.parametrize(
     "parameter_strategy",
     [
-        param(numerical_discrete_parameter(), id="NumericalDiscreteParameter"),
-        param(numerical_continuous_parameter(), id="NumericalContinuousParameter"),
-        param(categorical_parameter(), id="CategoricalParameter"),
-        param(task_parameter(), id="TaskParameter"),
-        param(custom_parameter(), id="CustomParameter"),
+        param(numerical_discrete_parameters(), id="NumericalDiscreteParameter"),
+        param(numerical_continuous_parameters(), id="NumericalContinuousParameter"),
+        param(categorical_parameters(), id="CategoricalParameter"),
+        param(task_parameters(), id="TaskParameter"),
+        param(custom_parameters(), id="CustomParameter"),
         param(
-            substance_parameter(),
+            substance_parameters(),
             id="SubstanceParameter",
             marks=pytest.mark.skipif(
                 not _CHEM_INSTALLED, reason="Optional chem dependency not installed."